requirements.txt
numpy~=1.26.4
pandas~=2.2.2
pillow~=10.4.0
geographiclib~=2.0
httpx~=0.27.0

aiofiles~=24.1.0
opencv-python~=4.10.0.84
python-dotenv~=1.0.1
pydantic~=2.8.2
streetview~=0.0.11
requests~=2.32.3
torch~=2.4.0
matplotlib~=3.9.2
ultralytics~=8.1.34
torchvision~=0.19.0
